feat: moved ensureBlock to core/ and small impr
