+++ lib/core/ensureBlock.js
@@ -1,5 +1,5 @@
 const ZMQClient = require('@dashevo/dashd-zmq');
-const waitForEvent = require('./waitForEvent');
+const waitForEvent = require('../util/waitForEvent');
 
 async function ensureBlock(socketClient, rpcClient, hash) {
   try {
