+++ lib/core/waitForCoreChainLockSyncFactory.js
@@ -1,7 +1,7 @@
 const { ChainLock } = require('@dashevo/dashcore-lib');
 const ZMQClient = require('@dashevo/dashd-zmq');
 const waitForEvent = require('../util/waitForEvent');
-const ensureBlock = require('../util/ensureBlock');
+const ensureBlock = require('./ensureBlock');
 
 /**
  * Check that core is synced (factory)
@@ -36,9 +36,9 @@ function waitForCoreChainLockSyncFactory(
 
     coreZMQClient.on(ZMQClient.TOPICS.rawchainlock, async (rawChainLockMessage) => {
       const socketChainLock = new ChainLock(rawChainLockMessage);
-      logger.info(`Received new chainlock ${socketChainLock.toString()}`);
       await ensureBlock(coreZMQClient, coreRpcClient, socketChainLock.blockHash);
       latestCoreChainLock.update(socketChainLock);
+      logger.debug(`Received new chainlock ${socketChainLock.toString()}`);
     });
 
     // Because a Chainlock may happen before its block, we also subscribe to rawblock
