+++ lib/util/waitForEvent.js
@@ -5,7 +5,7 @@
  * @param {String} [value] - Optionally will verify value to equal response
  * @return {Promise<any>} The promise to await on.
  */
-async function waitForEvent(emitter, eventName, value) {
+async function waitForEvent(emitter, eventName, value = undefined) {
   return new Promise((resolve) => {
     emitter.on(eventName, (response) => {
       if (value === undefined || value === response) {
